Corrected TDY Duration and improved display
TDY Duration was 1 days too few.  Changed calculation so that start date and end date both were included in calculation.  To do this, I changed the calculate_tdy_duration() function of class Bank to work on Integers instead of datetime objects.  This had the added bonus of changing the display from a datetime object to an integer, cleaning up the display.  I also added the word "days" after the TDY duration display.